refactor: 캡슐 관련 함수 분리
